clean-up data folder
