+++ whotracksme/data/__init__.py
@@ -1,6 +1,6 @@
 
-from whotracksme.data.db.trackerdb import load_tracker_db
 from whotracksme.data.loader import (
+    load_tracker_db,
     load_apps,
     load_overview,
     load_companies,
