+++ whotracksme/main.py
@@ -54,4 +54,3 @@ def main():
 
 if __name__ == '__main__':
     main()
-
